+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added check and exception for NaN data in the adjoint pipeline to raise issue to user before adjoint source creation failure.
 - Added autograd support for `TerminalComponentModeler` and `ModalComponentModeler`.
 - Added `initialize_params_from_simulation` to `tidy3d.plugins.autograd.invdes` to initialize topology design regions from an underlying simulation geometry.
+- Added autograd support for sidewall angles in `td.Cylinder` and `td.PolySlab`.
 
 ### Changed
 - Validate mode solver object for large number of grid points on the modal plane.